MAINT: Move PETSc solver interface to dedicated module
